+++ NEWS.md
@@ -1,5 +1,10 @@
 # memtools (development version)
 
+- Added `root_precious_list()` to retrieve the R precious list. This
+  requires a custom build of R where `R_PreciousList` is redefined as
+  a non-static symbol.
+
+
 # memtools 0.1.0
 
 Initial release